requirements.txt
absl-py==2.1.0
accelerate==1.0.1
ace_tools==0.0
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.12.0
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
Automat==22.10.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blinker==1.7.0
blis==0.7.11
boto3==1.34.131
botocore==1.34.131
cachetools==5.4.0
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloud-init==24.1.3
cloudpathlib==0.18.1
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
component==0.0.1
confection==0.1.5
configobj==5.0.8
constantly==23.10.4
contourpy==1.2.1
cryptography==41.0.7
cupy-cuda12x==12.3.0
cupy-wheel==12.3.0
cycler==0.12.1
cymem==2.0.8
cytoolz==0.12.3
dataclasses-json==0.6.7
datasets==2.20.0
dbus-python==1.3.2
debugpy==1.8.2
decorator==5.1.1
deeplake==3.9.26
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
distro-info==1.7+build1
dnspython==2.6.1
docopt==0.6.2
ec2-hibinit-agent==1.0.0
einops==0.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fastjsonschema==2.20.0
fastrlock==0.8.2
filelock==3.15.4
FlagEmbedding @ git+https://github.com/FlagOpen/FlagEmbedding.git@bda01e88b4cb03cf8251fdf9529fd1931daa91b3
flash-attn==2.6.3
flatbuffers==24.3.25
floret==0.10.5
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gpustat==1.1.1
greenlet==3.1.0
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
hibagent==1.0.1
httpcore==1.0.5
httplib2==0.20.4
httpx==0.27.0
huggingface-hub==0.24.2
humbug==0.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
InstructorEmbedding==1.0.1
interchange==2021.0.4
ipython==8.26.0
isoduration==20.11.0
jedi==0.19.1
jellyfish==1.0.4
Jinja2==3.1.2
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpickle==3.2.2
jsonpointer==2.0
jsonschema==4.10.3
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.11
kaleido==0.2.1
keras==3.4.1
kiwisolver==1.4.5
langchain==0.3.4
langchain-aws==0.2.2
langchain-community==0.3.1
langchain-core==0.3.13
langchain-text-splitters==0.3.0
langcodes==3.4.0
langsmith==0.1.129
language_data==1.2.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llama-cloud==0.1.4
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.20
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-postprocessor-flag-embedding-reranker==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.12
llvmlite==0.43.0
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==2.2.0
mistune==3.0.2
ml-dtypes==0.4.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.14.2
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
nltk==3.9.1
node2vec==0.5.0
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.52.2
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
pansi==2020.7.3
parso==0.8.4
pathos==0.3.3
peft==0.13.2
pexpect==4.9.0
pillow==10.4.0
pipx==1.4.3
platformdirs==4.2.2
plotly==5.24.1
pooch==1.8.2
portalocker==2.10.1
pox==0.3.5
ppft==1.7.6.9
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py2neo==2021.2.4
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.25.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
pydeck==0.9.1
pydotplus==2.0.2
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pymarc==5.2.2
pymongo==4.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==4.3.1
pyphen==0.15.0
pyrsistent==0.20.0
pyserial==3.5
pyspellchecker==0.8.1
python-apt==2.7.7+ubuntu1
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
s3transfer==0.10.2
sacrebleu==2.4.3
sacred==0.8.5
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
service-identity==24.1.0
setuptools==68.1.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sos==4.5.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
ssh-import-id==5.11
stack-data==0.6.3
streamlit==1.40.1
striprtf==0.0.26
sympy==1.13.1
systemd-python==235
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
terminado==0.18.1
textacy==0.13.0
tf_keras==2.17.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
Twisted==24.3.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.0.7
userpath==1.9.1
wadllib==1.3.6
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.42.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xkit==0.0.0
xxhash==3.4.1
yarl==1.9.4
zope.interface==6.1
